requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.3
sqlalchemy~=2.0.31
psycopg2-binary==2.9.9
pydantic~=2.8.2
requests~=2.32.3
alembic~=1.13.2
python-dotenv~=1.0.1
email-validator
pydantic-settings
python-jose
passlib~=1.7.4
passlib==1.7.4
bcrypt==3.2.0

httpx~=0.27.0
starlette~=0.37.2
PyJWT~=2.8.0
dnspython~=2.6.1
idna~=3.7
httpcore~=1.0.5
anyio~=4.4.0
h11~=0.14.0
pip~=24.2
filelock~=3.15.4
rsa~=4.9
pyasn1~=0.6.0
ecdsa~=0.19.0
Mako~=1.3.5
Pygments~=2.18.0
MarkupSafe~=2.1.5
typing_extensions~=4.12.2
rich~=13.7.1
PyYAML~=6.0.1
sniffio~=1.3.1
six~=1.16.0
click~=8.1.7
certifi~=2024.7.4
mdurl~=0.1.2
typer~=0.12.3
shellingham~=1.5.4
uvloop~=0.19.0
email_validator~=2.2.0
pydantic_core~=2.20.1
websockets~=12.0
urllib3~=2.2.2
httptools~=0.6.1
watchfiles~=0.22.0
psycopg2~=2.9.9
python-multipart~=0.0.9
pycparser~=2.22
Jinja2~=3.1.4
